+++ src/game/tests.rs
@@ -611,6 +611,7 @@ mod triumphs;
 mod turn_and_loyalty;
 mod tutors_and_fetch_lands;
 mod unblocked_attackers;
+mod unbounded_targets;
 mod unleash;
 mod untap_caps;
 mod untap_skips;